完成Role维护

+++ src/main/java/com/yfk/service/impl/RoleManagerImpl.java
@@ -9,6 +9,7 @@ import org.springframework.stereotype.Service;
 import com.yfk.model.PermissionType;
 import com.yfk.model.Role;
 import com.yfk.model.RolePermission;
+import com.yfk.model.UserRole;
 import com.yfk.service.RoleManager;
 import com.yfk.service.UniversalManager;
 import com.yfk.webapp.util.PrincipalNullException;
@@ -37,11 +38,27 @@ public class RoleManagerImpl extends GenericManagerImpl<Role, String>implements 
 			}
 		}
 	}
-	
+
+	@Override
+	public void saveRoleUser(String roleCode, List<String> assignedUsers) throws PrincipalNullException {
+		this.universalManager.executeByHql("delete from UserRole where roleCode = ?", new Object[] { roleCode });
+
+		if (!CollectionHelper.isEmpty(assignedUsers)) {
+			for (String assignedUser : assignedUsers) {
+				UserRole userRole = new UserRole();
+
+				userRole.setRoleCode(roleCode);
+				userRole.setUsername(assignedUser);
+
+				this.universalManager.save(userRole);
+			}
+		}
+
+	}
+
 	public void deleteRole(String roleCode) {
-		this.universalManager.executeByHql("delete from RolePermission where roleCode = ?",
-				new Object[] { roleCode });
-		
+		this.universalManager.executeByHql("delete from RolePermission where roleCode = ?", new Object[] { roleCode });
+
 		this.universalManager.remove(Role.class, roleCode);
 	}
 }
